classifier train and diffusion train successful

+++ src/tabular_synthesis/synthesizer/model/guided_diffusion/gaussian_diffusion.py
@@ -757,7 +757,7 @@ class GaussianDiffusion:
         if model_kwargs is None:
             model_kwargs = {}
         if noise is None:
-            noise = th.randn_like(x_start)
+            noise = th.randn_like(x_start.float()) # <--- evtl. .float() löschen auf gpu
         x_t = self.q_sample(x_start, t, noise=noise)
 
         terms = {}
